src/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
exceptiongroup==1.2.0
Faker==12.0.1
ffmpeg-python==0.2.0
future==0.18.3
gunicorn==21.2.0
iniconfig==2.0.0
kombu==5.3.5
Markdown==3.5.2
mixer==7.2.2
mock==5.1.0
packaging==23.2
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pytest==8.0.1
pytest-celery==0.0.0
pytest-django==4.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
